+++ src/vector.h
@@ -9,33 +9,33 @@
 
 #define declareVector(VectorTypeName, TValue, methodPrefix) \
 typedef struct VectorTypeName##Struct *VectorTypeName; \
-VectorTypeName make##VectorTypeName(const long int capacity); \
-void initialize##VectorTypeName(VectorTypeName vector, const long int capacity); \
+VectorTypeName make##VectorTypeName(const int32_t capacity); \
+void initialize##VectorTypeName(VectorTypeName vector, const int32_t capacity); \
 void destroy##VectorTypeName(VectorTypeName vector); \
 void free##VectorTypeName(VectorTypeName vector); \
 void methodPrefix##Push(VectorTypeName vector, const TValue value); \
 void methodPrefix##Clear(VectorTypeName vector); \
 TValue* methodPrefix##Begin(VectorTypeName vector); \
 TValue* methodPrefix##End(VectorTypeName vector); \
-long int methodPrefix##Size(VectorTypeName vector); \
+int32_t methodPrefix##Size(VectorTypeName vector); \
 
 
 #define defineVectorStruct(VectorTypeName, TValue, methodPrefix) \
 typedef struct VectorTypeName##Struct { \
 	TValue *values; \
-	long int size; \
-	long int capacity; \
+	int32_t size; \
+	int32_t capacity; \
 } VectorTypeName##Struct; \
 
 
 #define defineVector(VectorTypeName, TValue, methodPrefix) \
-VectorTypeName make##VectorTypeName(const long int capacity) { \
+VectorTypeName make##VectorTypeName(const int32_t capacity) { \
 	VectorTypeName vector = malloc(sizeof(VectorTypeName)); \
 	assert(vector); \
 	initialize##VectorTypeName(vector, capacity); \
 	return vector; \
 } \
-void initialize##VectorTypeName(VectorTypeName vector, const long int capacity) { \
+void initialize##VectorTypeName(VectorTypeName vector, const int32_t capacity) { \
 	vector->values = malloc(sizeof(TValue) * capacity); \
 	assert(vector->values); \
     vector->size = 0; \
@@ -67,7 +67,7 @@ TValue* methodPrefix##Begin(VectorTypeName vector) { \
 TValue* methodPrefix##End(VectorTypeName vector) { \
     return vector->values + vector->size; \
 } \
-long int methodPrefix##Size(VectorTypeName vector) { \
+int32_t methodPrefix##Size(VectorTypeName vector) { \
 	return vector->size; \
 } \
